wandb/run-20231118_180856-m7z0zv42/files/requirements.txt
appdirs==1.4.4
bottleneck==1.3.5
brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
contourpy==1.0.5
cryptography==41.0.3
cycler==0.11.0
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
filelock==3.9.0
fonttools==4.25.0
fsspec==2023.10.0
gitdb==4.0.11
gitpython==3.1.40
gudhi==3.8.0
huggingface-hub==0.19.4
idna==3.4
importlib-resources==6.1.0
jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
markupsafe==2.1.1
matplotlib==3.8.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
munch==4.0.0
munkres==1.1.4
networkx==3.1
numexpr==2.8.7
numpy==1.26.0
packaging==23.1
pandas==2.1.1
pathtools==0.1.2
pillow==10.0.1
pip==23.3
piq==0.8.0
ply==3.11
positional-encodings==6.0.1
pretrainedmodels==0.7.4
protobuf==4.24.4
psutil==5.9.6
pycparser==2.21
pyopenssl==23.2.0
pyparsing==3.0.9
pyqt5-sip==12.13.0
pyqt5==5.15.10
pysocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyyaml==6.0.1
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.0
scipy==1.11.3
segmentation-models-pytorch==0.3.3
sentry-sdk==1.32.0
setproctitle==1.3.3
setuptools==68.0.0
sip==6.7.12
six==1.16.0
smmap==5.0.1
sympy==1.11.1
threadpoolctl==2.2.0
tifffile==2023.9.26
timm==0.9.2
tomli==2.0.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
typing-extensions==4.7.1
tzdata==2023.3
urllib3==1.26.18
wandb==0.15.12
warmup-scheduler==0.3.2
wheel==0.41.2
win-inet-pton==1.1.0
zipp==3.11.0